Refs #21039 -- Added support for concurrent indexes to migrations
The previous commit already added support for concurrent creation to
models.Index(). However, native support in the migration framework, to
properly split non-atomic or "standalone" operations into dedicated
migrations was not around yet. This commit will treat operations
signalling that they are "__standalone__" that they need to have their
own migration and cannot be pushed in a previous migration.

I opted for individual migrations for multiple concurrent indexes to
reduce the risk of a migration failing half-way, leaving the database in
an inconsistent state.